requirements.txt
opencv-python
ultralytics
Pillow
numpy
